refactor: update code snippets to TS + latest SC versions
