+++ fetch-profile/fetch-json-data.ts
@@ -1,5 +1,5 @@
 import { ERC725 } from '@erc725/erc725.js';
-import lsp3ProfileSchema from '@erc725/erc725.js/schemas/LSP3ProfileMetadata.json' assert { type: 'json' };
+import lsp3ProfileSchema from '@erc725/erc725.js/schemas/LSP3ProfileMetadata.json';
 
 // Initatiate erc725.js
 const erc725js = new ERC725(
