--- key-manager/3rd-party-permissions.js
@@ -1,1 +1,0 @@
-// WIP
