--- key-manager/constants.js
@@ -1,30 +1,0 @@
-// keccak256('AddressPermissions[]')
-const PERMISSIONS_ARRAY =
-  '0xdf30dba06db6a30e65354d9a64c609861f089545ca58c6b4dbe31a5f338cb0e3';
-
-// prettier-ignore
-const ADDRESSES = {
-  PERMISSIONS:      "0x4b80742de2bf82acb3630000", // AddressPermissions:Permissions:<address> --> bytes32
-  ALLOWEDADDRESSES: "0x4b80742de2bfc6dd6b3c0000", // AddressPermissions:AllowedAddresses:<address> --> address[]
-  ALLOWEDFUNCTIONS: "0x4b80742de2bf8efea1e80000", // AddressPermissions:AllowedFunctions:<address> --> bytes4[]
-}
-
-// prettier-ignore
-const PERMISSIONS = {
-  CHANGEOWNER:      "0x0000000000000000000000000000000000000000000000000000000000000001", // 0000 0000 0000 0001
-  CHANGEPERMISSIONS:"0x0000000000000000000000000000000000000000000000000000000000000002", // .... .... .... 0010
-  ADDPERMISSIONS:   "0x0000000000000000000000000000000000000000000000000000000000000004", // .... .... .... 0100
-  SETDATA:          "0x0000000000000000000000000000000000000000000000000000000000000008", // .... .... .... 1000
-  CALL:             "0x0000000000000000000000000000000000000000000000000000000000000010", // .... .... 0001 ....
-  STATICCALL:       "0x0000000000000000000000000000000000000000000000000000000000000020", // .... .... 0010 ....
-  DELEGATECALL:     "0x0000000000000000000000000000000000000000000000000000000000000040", // .... .... 0100 ....
-  DEPLOY:           "0x0000000000000000000000000000000000000000000000000000000000000080", // .... .... 1000 ....
-  TRANSFERVALUE:    "0x0000000000000000000000000000000000000000000000000000000000000100", // .... 0001 .... ....
-  SIGN:             "0x0000000000000000000000000000000000000000000000000000000000000200", // .... 0010 .... ....
-}
-
-module.exports = {
-  ADDRESSES,
-  PERMISSIONS,
-  PERMISSIONS_ARRAY,
-};
