--- key-manager/get-controller-keys.js
@@ -1,1 +1,0 @@
-// WIP
