+++ key-manager/set-permissions.ts
@@ -1,17 +1,19 @@
 // Imports
 import { ethers } from 'ethers';
-import UniversalProfile from '@lukso/lsp-smart-contracts/artifacts/UniversalProfile.json' assert { type: 'json' };
-import KeyManager from '@lukso/lsp-smart-contracts/artifacts/LSP6KeyManager.json' assert { type: 'json' };
-import { ADDRESSES, PERMISSIONS, PERMISSIONS_ARRAY } from './constants';
+import {
+  PERMISSIONS,
+  ERC725YDataKeys,
+} from '@lukso/lsp-smart-contracts/constants';
+import UniversalProfile from '@lukso/lsp-smart-contracts/artifacts/UniversalProfile.json';
 
 // Connect to the LUKSO Testnet
 const RPC_ENDPOINT = 'https://rpc.testnet.lukso.network';
 const provider = new ethers.providers.JsonRpcProvider(RPC_ENDPOINT);
 
-const myUniversalProfileAddress = '0x...';
-const myKeyManagerAddress = '0x...';
+const myUniversalProfileAddress = '0x9fc7e5095A054dfA3c6b237E0e5d686638394248';
+const myKeyManagerAddress = '0x87fa9105cA247897Acb4F12Ddf6EC3CEF23F6059';
 
-const walletPrivateKey = '<your-wallet-private-key>'; // Replace with your wallet's private key
+const walletPrivateKey = '0x...'; // Replace with your wallet's private key
 const wallet = new ethers.Wallet(walletPrivateKey, provider);
 
 const universalProfile = new ethers.Contract(
@@ -21,16 +23,18 @@ const universalProfile = new ethers.Contract(
 );
 
 // EOA address of an exemplary person
-let bobAddress = '0x...';
+let bobAddress = '0xcafecafecafecafecafecafecafecafecafecafe';
 let bobPermissions = PERMISSIONS.SETDATA;
 
 // give the permission SETDATA to Bob
 async function setPermission() {
   try {
     const permissionData = [
-      ADDRESSES.PERMISSIONS + bobAddress.substring(2), // allow Bob to setData on your UP
-      PERMISSIONS_ARRAY, // length of AddressPermissions[]
-      PERMISSIONS_ARRAY.slice(0, 34) + '00000000000000000000000000000001', // add Bob's address into the list of permissions
+      ERC725YDataKeys.LSP6['AddressPermissions:Permissions'] +
+        bobAddress.substring(2), // allow Bob to setData on your UP
+      ERC725YDataKeys.LSP6['AddressPermissions[]'].length, // length of AddressPermissions[]
+      ERC725YDataKeys.LSP6['AddressPermissions[]'].index +
+        '00000000000000000000000000000001', // add Bob's address into the list of permissions
     ];
 
     const permissionParams = [
@@ -39,9 +43,10 @@ async function setPermission() {
       bobAddress,
     ];
 
-    const tx = await universalProfile.populateTransaction[
-      'setData(bytes32[],bytes[])'
-    ](permissionData, permissionParams);
+    const tx = await universalProfile.populateTransaction.setDataBatch(
+      permissionData,
+      permissionParams,
+    );
 
     const txResponse = await wallet.sendTransaction({
       to: myKeyManagerAddress,
