+++ metadata-detection/digital-asset-check.ts
@@ -1,24 +1,24 @@
 import { ERC725 } from '@erc725/erc725.js';
-import lsp9VaultSchema from '@erc725/erc725.js/schemas/LSP9Vault.json' assert { type: 'json' };
+import lsp4Schema from '@erc725/erc725.js/schemas/LSP4DigitalAsset.json';
 
 // Initatiate erc725.js
 const erc725js = new ERC725(
-  lsp9VaultSchema,
-  '0x9139def55c73c12bcda9c44f12326686e3948634',
+  lsp4Schema,
+  '0x6395b330F063F96579aA8F7b59f2584fb9b6c3a5',
   'https://rpc.testnet.lukso.gateway.fm',
   {},
 );
 
 // Fetch the supported storage standard
-let isLSP9 = false;
+let isLSP4 = false;
 
 // Verify if the standard is supported (value !== null)
-const data = await erc725js.getData('SupportedStandards:LSP9Vault');
+const data = await erc725js.getData('SupportedStandards:LSP4DigitalAsset');
 if (data.value !== null) {
-  isLSP9 = true;
+  isLSP4 = true;
 }
 
-console.log(isLSP9);
+console.log(isLSP4);
 
 /*
 Supported schemas from erc725.js library:
