+++ transfer-token/backend-token-transaction.ts
@@ -1,7 +1,7 @@
 // Imports
 import { ethers } from 'ethers';
-import LSP7Mintable from '@lukso/lsp-smart-contracts/artifacts/LSP7Mintable.json' assert { type: 'json' };
-import UniversalProfile from '@lukso/lsp-smart-contracts/artifacts/UniversalProfile.json' assert { type: 'json' };
+import LSP7Mintable from '@lukso/lsp-smart-contracts/artifacts/LSP7Mintable.json';
+import UniversalProfile from '@lukso/lsp-smart-contracts/artifacts/UniversalProfile.json';
 
 // Connect to the mainnet or testnet
 const RPC_ENDPOINT = 'https://rpc.testnet.lukso.gateway.fm';
