+++ update-profile/sample-metadata.json
@@ -13,18 +13,22 @@
       {
         "width": 640,
         "height": 609,
-        "hashFunction": "keccak256(bytes)",
-        "hash": "0xe459e5769af85b09fb43bb8eaac561e196d58c0f5da3c5e150b6695898089402",
-        "url": "ipfs://QmPCz896rcZmq8F3FuUkJinRUmPgnZGjPvZL71nAaL7Fsx"
+        "url": "ipfs://QmPCz896rcZmq8F3FuUkJinRUmPgnZGjPvZL71nAaL7Fsx",
+        "verification": {
+          "method": "keccak256(bytes)",
+          "data": "0xe459e5769af85b09fb43bb8eaac561e196d58c0f5da3c5e150b6695898089402" // bytes32 hash of the image
+        }
       }
     ],
     "backgroundImage": [
       {
         "width": 1024,
         "height": 576,
-        "hashFunction": "keccak256(bytes)",
-        "hash": "0x1c19780d377a7b01f7dcf16e0ebffd225e29d2235625009f67cf9d86a32a79e1",
-        "url": "ipfs://QmPMmEpKnmgACsWjhDUheF8TEKpspzQhAkjbY4EBbR4jgP"
+        "url": "ipfs://QmPMmEpKnmgACsWjhDUheF8TEKpspzQhAkjbY4EBbR4jgP",
+        "verification": {
+          "method": "keccak256(bytes)",
+          "data": "0x1c19780d377a7b01f7dcf16e0ebffd225e29d2235625009f67cf9d86a32a79e1" // bytes32 hash of the image
+        }
       }
     ]
   }
